pyproject.toml
[tool.poetry]
name = "repl-nix-notboringruns"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.39.0"
psycopg2-binary = "^2.9.9"
folium = "^0.17.0"
streamlit-folium = "^0.23.0"
requests = "^2.32.3"
stravalib = "^2.0"
flask = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
